*7207* Deal with .. filename case

+++ lib/tinymce/jscripts/tiny_mce/plugins/jbimages/ci/application/controllers/editor.php
@@ -197,7 +197,7 @@ class Editor extends CI_Controller {
 			// Make sure image exists in upload path
 			$imagePath = $imageDir . '/' . basename($imageName);
 
-			if(file_exists($imagePath)) {
+			if(is_file($imagePath) && file_exists($imagePath)) {
 				unlink($imagePath);
 			} else {
 				show_error('File does not exist');
